+++ lib/ts/recipe/webauthn/constants.ts
@@ -15,3 +15,5 @@
 export const DEFAULT_WEBAUTHN_RECOVERY_PATH = "/webauthn/recover";
 
 export const DEFAULT_WEBAUTHN_SEND_RECOVERY_EMAIL_PATH = "/webauthn/recover/send-email";
+
+export const DEFAULT_WEBAUTHN_MFA_PATH = "/mfa/webauthn";