+++ lib/ts/recipe/webauthn/recipe.tsx
@@ -14,10 +14,13 @@
  */
 
 import WebauthnWebJS from "supertokens-web-js/lib/build/recipe/webauthn";
+import { PostSuperTokensInitCallbacks } from "supertokens-web-js/utils/postSuperTokensInitCallbacks";
 
+import PasskeyIcon from "../../components/assets/passkeyIcon";
 import { SSR_ERROR } from "../../constants";
 import { getDefaultRedirectionURLForPath, isTest } from "../../utils";
 import AuthRecipe from "../authRecipe";
+import MultiFactorAuth from "../multifactorauth/recipe";
 import { FactorIds } from "../multifactorauth/types";
 
 import { DEFAULT_WEBAUTHN_SEND_RECOVERY_EMAIL_PATH } from "./constants";
@@ -39,6 +42,14 @@ import type {
 } from "../../types";
 import type RecipeModule from "../recipeModule";
 
+export const webauthnFactor = {
+    id: FactorIds.WEBAUTHN,
+    name: "WEBAUTHN_MFA_NAME",
+    description: "WEBAUTHN_MFA_DESCRIPTION",
+    path: "/mfa/webauthn",
+    logo: PasskeyIcon,
+};
+
 export default class Webauthn extends AuthRecipe<
     GetRedirectionURLContext,
     PreAndPostAPIHookAction,
@@ -57,8 +68,12 @@ export default class Webauthn extends AuthRecipe<
     ) {
         super(config);
 
-        // We can ideally call postInitCallbacks to set MFA's if
-        // we are using it.
+        PostSuperTokensInitCallbacks.addPostInitCallback(() => {
+            const mfa = MultiFactorAuth.getInstance();
+            if (mfa !== undefined) {
+                mfa.addMFAFactors([webauthnFactor]);
+            }
+        });
     }
 
     public getFirstFactorsForAuthPage(): string[] {