+++ lib/ts/recipe/webauthn/types.ts
@@ -29,6 +29,12 @@ import type {
     Config as AuthRecipeModuleConfig,
 } from "../authRecipe/types";
 import type { ContinueWithPasskeyWithOverride } from "./components/themes/continueWithPasskey";
+import type {
+    WebauthnMFASignIn,
+    WebauthnMFALoadingScreen,
+    WebauthnMFASignUp,
+    WebauthnMFASignUpConfirmation,
+} from "./components/themes/mfa";
 import type { PasskeyRecoveryEmailSent } from "./components/themes/sendRecoveryEmail/emailSent";
 import type {
     WebauthnRecoverAccount,
@@ -39,6 +45,8 @@ import type { ContinueWithoutPasskey } from "./components/themes/signUp/continue
 import type { PasskeyFeatureBlock } from "./components/themes/signUp/featureBlocks";
 import type { SignUpFormInner } from "./components/themes/signUp/signUpForm";
 import type { SignUpSomethingWentWrong } from "./components/themes/signUp/somethingWentWrong";
+import type { Dispatch } from "react";
+import type React from "react";
 import type { RecipeInterface } from "supertokens-web-js/recipe/webauthn";
 import type WebJSRecipe from "supertokens-web-js/recipe/webauthn";
 import type { User } from "supertokens-web-js/types";
@@ -124,6 +132,10 @@ export type ComponentOverrideMap = {
     WebauthnContinueWithoutPasskey_Override?: ComponentOverride<typeof ContinueWithoutPasskey>;
     WebauthnPasskeySignUpForm_Override?: ComponentOverride<typeof SignUpFormInner>;
     WebauthnPasskeySignUpSomethingWentWrong_Override?: ComponentOverride<typeof SignUpSomethingWentWrong>;
+    WebauthnMFASignIn_Override?: ComponentOverride<typeof WebauthnMFASignIn>;
+    WebauthnMFALoadingScreen_Override?: ComponentOverride<typeof WebauthnMFALoadingScreen>;
+    WebauthnMFASignUp_Override?: ComponentOverride<typeof WebauthnMFASignUp>;
+    WebauthnMFASignUpConfirmation_Override?: ComponentOverride<typeof WebauthnMFASignUpConfirmation>;
 };
 
 export type SignUpThemeBaseProps = {
@@ -248,4 +260,57 @@ export type ContinueWithPasskeyProps = {
 export type EmailSentProps = {
     email: string;
     onEmailChangeClick: () => void;
+};
+
+export type WebAuthnMFAAction =
+    | {
+          type: "setError";
+          accessDenied?: boolean;
+          error: string | undefined;
+      }
+    | {
+          type: "load";
+          deviceSupported: boolean;
+          error: string | undefined;
+          email: string | undefined;
+          showBackButton: boolean;
+          canRegisterPasskey: boolean;
+          hasRegisteredPassKey: boolean;
+      };
+
+type WebAuthnMFAInitialState = {
+    error: undefined;
+    loaded: false;
+    accessDenied: boolean;
+    deviceSupported: boolean;
+    showBackButton: boolean;
+    canRegisterPasskey: false;
+    hasRegisteredPassKey: false;
+    email: undefined;
+};
+
+type WebAuthnMFALoadedState = {
+    error: string | undefined;
+    loaded: true;
+    accessDenied: boolean;
+    deviceSupported: boolean;
+    showBackButton: boolean;
+    canRegisterPasskey: boolean;
+    hasRegisteredPassKey: boolean;
+    email: string | undefined;
+};
+
+export type WebAuthnMFAState = WebAuthnMFAInitialState | WebAuthnMFALoadedState;
+
+export type WebAuthnMFAProps = {
+    recipeImplementation: RecipeImplementation;
+    config: NormalisedConfig;
+    onBackButtonClicked: () => void;
+    onSignOutClicked: () => void;
+    onSignIn: () => Promise<void>;
+    onSignUp: (email: string) => Promise<void>;
+    onRecoverAccountClick: () => void;
+    userContext?: UserContext;
+    featureState: WebAuthnMFAState;
+    dispatch: Dispatch<WebAuthnMFAAction>;
 };